Refactor BOM Tables to use a common partial
Created a single BOM table partial to use for the instruments,
handvalves, and equipment BOM tables.  These tables currently have their own independently formatted tables.  Switching to a single partial reduces the amount of CSS required, and also ensures a consistent format.  Using a single partial required adding two methods to each model - one to list the data_attributes that we want to report, and one to create a hash that contained the attribute, the icon we want to use for the headers, and that contained the text of the title we want to use for each header.

As a result of these new methods, it's expected that the properties
partial for each model could be combined into a single shared partial as
well.